fix: import errors, refactor packages
