+++ src/presentation/commands/chatCommands/sendMessage.ts
@@ -2,7 +2,7 @@ import { Server } from 'socket.io'
 import { Ack } from '../../../application/message'
 import { SessionCommandHandlers } from '../../../application/commands/sessionCommandHandlers'
 import { SessionNotifications } from '../../notifications/sessionNotifications'
-import { SendMessageCommand } from '../../../application/commands/sessionCommands'
+import { SendMessageCommand } from '../../../application/commands/chatCommands'
 
 /**
  * Send message command.
