+++ src/presentation/commands/videoCommands/videoCommands.ts
@@ -2,7 +2,7 @@ import { Server } from 'socket.io'
 import { Ack } from '../../../application/message'
 import { SessionCommandHandlers } from '../../../application/commands/sessionCommandHandlers'
 import { SessionNotifications } from '../../notifications/sessionNotifications'
-import { PlayVideoCommand, StopVideoCommand } from '../../../application/commands/sessionCommands'
+import { PlayVideoCommand, StopVideoCommand } from '../../../application/commands/videoCommands'
 
 /**
  * Play video command
